Create Lab 6 _1_4.sql
TASK
AdventureWorks is interested in finding out which products are costing more than they're being sold for, on average.
NOTE
Filter the query for the previous exercise to include only products where the cost is higher than the average selling price. Make sure to use the aliases provided, and default column names elsewhere.