Create Retrieving Product Information (2)
TASK
You are only interested in products which have a listed color in the database.
NOTE
Create a table variable and populate it with a list of distinct colors from the SalesLT.Product table. Then use the table variable to filter a query that returns the product ID, name, and color from the SalesLT.Product table so that only products with a color listed in the table variable are returned. You'll need to use NVARCHAR(15) in your solution and make sure to use the aliases provided.